Mage 3 Port + Vizier stuff

+++ code/modules/spells/spell_types/classunique/vizier/reversion.dm
@@ -1,14 +1,15 @@
 // Reversion - Origin Magic (Vizier)
 // The Vizier marks an adjacent ally or themselves, snapshotting their state and position.
 // Leaves a mark on the ground - that they can then activate at will to teleport back
-// It expires after 15 seconds
+// It expires after 25 seconds
 
-#define REVERSION_MARK_DURATION (15 SECONDS)
+#define REVERSION_MARK_DURATION (25 SECONDS)
 
-/datum/action/cooldown/spell/reversion
+/datum/action/cooldown/spell/vizier/reversion
 	button_icon = 'icons/mob/actions/classuniquespells/vizier.dmi'
 	name = "Reversion"
-	desc = "Marks an adjacent ally's body and position, granting them the ability to revert to their marked state within 15 seconds. The target chooses when to activate the revert. If the target is under Convergence when activating this, all healing gained after application will remain."
+	desc = "Marks an adjacent ally's body and position for 25 seconds, allowing them to return to their marked state at will."
+	fluff_desc = "Among the most demanding applications of Origin Magick, this art does not merely restore a prior state. It preserves one. For a fleeting moment, a Vizier anchors a person's place within the tapestry of time, allowing it to retrace its own history and reclaim a body, position, and condition once held."
 	button_icon_state = "reversion"
 	sound = 'sound/magic/timeforward.ogg'
 	spell_color = GLOW_COLOR_ARCANE
@@ -27,7 +28,7 @@
 
 	charge_required = TRUE
 	charge_time = 0.5 SECONDS
-	charge_drain = 0
+	hold_drain = 0
 	charge_slowdown = CHARGING_SLOWDOWN_SMALL
 	charge_sound = 'sound/magic/charging.ogg'
 	cooldown_time = 60 SECONDS
@@ -38,10 +39,9 @@
 
 	spell_requirements = SPELL_REQUIRES_NO_ANTIMAGIC | SPELL_REQUIRES_HUMAN | SPELL_REQUIRES_SAME_Z
 
-	/// Fatigue/mana cost for the Vizier's origin magic system.
-	var/cost = 3
+	cost = 3
 
-/datum/action/cooldown/spell/reversion/is_valid_target(atom/cast_on)
+/datum/action/cooldown/spell/vizier/reversion/is_valid_target(atom/cast_on)
 	. = ..()
 	if(!.)
 		return FALSE
@@ -55,7 +55,7 @@
 		return FALSE
 	return TRUE
 
-/datum/action/cooldown/spell/reversion/cast(atom/cast_on)
+/datum/action/cooldown/spell/vizier/reversion/cast(atom/cast_on)
 	. = ..()
 	var/mob/living/carbon/human/H = owner
 	if(!istype(H))
@@ -66,7 +66,7 @@
 		return FALSE
 
 	// Snapshot the target's current state
-	var/datum/action/cooldown/spell/reversion_trigger/trigger = new
+	var/datum/action/cooldown/spell/vizier/reversion_trigger/trigger = new
 	trigger.origin = get_turf(target)
 	trigger.brute = target.getBruteLoss()
 	trigger.burn = target.getFireLoss()
@@ -92,20 +92,23 @@
 	// Audio + visual feedback on the target
 	playsound(target.loc, 'sound/magic/timeforward.ogg', 50, FALSE)
 	target.apply_status_effect(/datum/status_effect/buff/reversion)
+	target.balloon_alert(target, "marked for reversion")
+	if(target != H)
+		target.balloon_alert(H, "marked for reversion")
 
 	// Feedback
 	if(target == H)
-		to_chat(H, span_notice("I mark myself for reversion. I can activate the revert at will."))
+		to_chat(H, span_purple("I mark myself for reversion. I can activate the revert at will."))
 	else
-		to_chat(target, span_warning("I feel a part of me was left behind... I can choose to revert back."))
-		to_chat(H, span_notice("I mark [target] for reversion. They can activate the revert at will."))
+		to_chat(target, span_purple("I feel a part of me was left behind... I can choose to revert back."))
+		to_chat(H, span_purple("I mark [target] for reversion. They can activate the revert at will."))
 
 	return TRUE
 
-/datum/action/cooldown/spell/reversion_trigger
+/datum/action/cooldown/spell/vizier/reversion_trigger
 	button_icon = 'icons/mob/actions/classuniquespells/vizier.dmi'
 	name = "Revert"
-	desc = "Activate to snap back to your marked position and restore your state. If you are under Convergence when this is activated, you retain any healing gained during the effect."
+	desc = "Activate to snap back to your marked position and restore your state."
 	button_icon_state = "reversion"
 	sound = 'sound/magic/timereverse.ogg'
 	spell_color = GLOW_COLOR_ARCANE
@@ -120,8 +123,8 @@
 	invocation_type = INVOCATION_NONE
 
 	charge_required = TRUE
-	charge_time = 0.5 SECONDS
-	charge_drain = 0
+	charge_time = 0 // Instant cuz do after already exists
+	hold_drain = 0
 	charge_slowdown = CHARGING_SLOWDOWN_SMALL
 	charge_sound = 'sound/magic/charging.ogg'
 	cooldown_time = 0
@@ -149,8 +152,10 @@
 	var/expiry_timer
 	/// Ground marker effect at the origin point.
 	var/obj/effect/reversion_marker/ground_marker
+	/// TRUE while the revert channel is running - blocks re-entry and marks us as committed.
+	var/reverting = FALSE
 
-/datum/action/cooldown/spell/reversion_trigger/Grant(mob/grant_to)
+/datum/action/cooldown/spell/vizier/reversion_trigger/Grant(mob/grant_to)
 	. = ..()
 	if(!owner)
 		return
@@ -160,11 +165,43 @@
 	if(origin)
 		ground_marker = new(origin)
 
-/datum/action/cooldown/spell/reversion_trigger/cast(atom/cast_on)
+/datum/action/cooldown/spell/vizier/reversion_trigger/cast(atom/cast_on)
 	. = ..()
 	var/mob/living/carbon/target = owner
 	if(!istype(target))
 		return FALSE
+
+	if(reverting)
+		return FALSE
+	reverting = TRUE
+
+	if(expiry_timer)
+		deltimer(expiry_timer)
+		expiry_timer = null
+
+
+	target.balloon_alert(target, "reverting...")
+	target.visible_message(span_purple("[target]'s body begins to flicker, slipping out of the present moment..."))
+	var/datum/progressbar/progbar = new(target, 2 SECONDS, target)
+	var/endtime = world.time + 2 SECONDS
+	var/starttime = world.time
+	while(world.time < endtime)
+		stoplag(1)
+		if(QDELETED(src) || QDELETED(target))
+			break
+		progbar.update(world.time - starttime)
+	qdel(progbar)
+
+
+	if(QDELETED(src) || QDELETED(target) || target != owner)
+		reverting = FALSE
+		return FALSE
+
+	var/turf/departure = get_turf(target)
+	target.visible_message(span_purple("[target] flickers and warps away, snapping backwards through time!"), span_purple("Time reverses - my body snaps back!"))
+	if(departure)
+		departure.balloon_alert_to_viewers("warps away!")
+		playsound(departure, 'sound/magic/timereverse.ogg', 100, FALSE)
 
 	// Teleport back
 	do_teleport(target, origin, no_effects = TRUE)
@@ -173,27 +210,14 @@
 	var/brutenew = target.getBruteLoss()
 	var/burnnew = target.getFireLoss()
 	var/oxynew = target.getOxyLoss()
-	var/toxinnew = target.getToxLoss()
 	target.adjust_fire_stacks(firestacks)
 	target.adjust_fire_stacks(sunderfirestacks, /datum/status_effect/fire_handler/fire_stacks/sunder)
 	target.adjust_fire_stacks(divinefirestacks, /datum/status_effect/fire_handler/fire_stacks/divine)
-	if(target.has_status_effect(/datum/status_effect/buff/convergence))
-		if(brutenew>brute)
-			target.adjustBruteLoss(brutenew * -1 + brute)
-		if(burnnew>burn)
-			target.adjustFireLoss(burnnew * -1 + burn)
-		if(oxynew>oxy)
-			target.adjustOxyLoss(oxynew * -1 + oxy)
-		if(toxinnew>toxin)
-			target.adjustToxLoss(target.getToxLoss() * -1 + toxin)
-		if(target.blood_volume<blood)
-			target.blood_volume = blood
-	else
-		target.adjustBruteLoss(brutenew * -1 + brute)
-		target.adjustFireLoss(burnnew * -1 + burn)
-		target.adjustOxyLoss(oxynew * -1 + oxy)
-		target.adjustToxLoss(target.getToxLoss() * -1 + toxin)
-		target.blood_volume = blood
+	target.adjustBruteLoss(brutenew * -1 + brute)
+	target.adjustFireLoss(burnnew * -1 + burn)
+	target.adjustOxyLoss(oxynew * -1 + oxy)
+	target.adjustToxLoss(target.getToxLoss() * -1 + toxin)
+	target.blood_volume = blood
 
 	// Remove any wounds gained after the mark
 	for(var/datum/wound/wound as anything in target.get_wounds())
@@ -204,8 +228,9 @@
 		else
 			target.simple_remove_wound(wound)
 
+	// Announce the arrival at the return point
 	playsound(target.loc, 'sound/magic/timereverse.ogg', 100, FALSE)
-	to_chat(target, span_warning("Time reverses - my body snaps back!"))
+	target.balloon_alert_to_viewers("snaps into place!", "I snap back!")
 
 	// Remove the status effect and clean up
 	target.remove_status_effect(/datum/status_effect/buff/reversion)
@@ -213,17 +238,20 @@
 	return TRUE
 
 /// Called when the mark expires without being used.
-/datum/action/cooldown/spell/reversion_trigger/proc/expire()
+/datum/action/cooldown/spell/vizier/reversion_trigger/proc/expire()
+	// A revert is already committed and channeling - let it finish, don't tear down under it.
+	if(reverting)
+		return
 	if(!owner)
 		qdel(src)
 		return
-	to_chat(owner, span_notice("The reversion mark fades."))
+	to_chat(owner, span_purple("The reversion mark fades."))
 	var/mob/living/L = owner
 	L.remove_status_effect(/datum/status_effect/buff/reversion)
 	cleanup()
 
 /// Remove this spell from the target and delete it.
-/datum/action/cooldown/spell/reversion_trigger/proc/cleanup()
+/datum/action/cooldown/spell/vizier/reversion_trigger/proc/cleanup()
 	if(expiry_timer)
 		deltimer(expiry_timer)
 		expiry_timer = null
@@ -232,7 +260,7 @@
 		Remove(owner)
 	qdel(src)
 
-/datum/action/cooldown/spell/reversion_trigger/Destroy()
+/datum/action/cooldown/spell/vizier/reversion_trigger/Destroy()
 	if(expiry_timer)
 		deltimer(expiry_timer)
 		expiry_timer = null
@@ -258,42 +286,3 @@
 	mouse_opacity = MOUSE_OPACITY_TRANSPARENT
 
 #undef REVERSION_MARK_DURATION
-
-//Caustic Edit - Re-Add Convergence here!
-/obj/effect/proc_holder/spell/invoked/convergence
-	name = "Convergence"
-	desc = "Converges the targets past and present, causing them to heal 50% more."
-	overlay_state = "convergence"
-	releasedrain = 30
-	chargedrain = 0
-	chargetime = 0
-	range = 4
-	warnie = "sydwarning"
-	movement_interrupt = FALSE
-	req_items = list(/obj/item/clothing/neck/roguetown/psicross)
-	sound = list('sound/magic/convergence1.ogg','sound/magic/convergence2.ogg','sound/magic/convergence3.ogg','sound/magic/convergence4.ogg')
-	action_icon = 'icons/mob/actions/classuniquespells/vizier.dmi'
-	invocation_type = "none"
-	associated_skill = /datum/skill/magic/arcane
-	antimagic_allowed = TRUE
-	recharge_time = 30 SECONDS
-	miracle = FALSE
-	cost = 4
-	devotion_cost = 0
-
-/obj/effect/proc_holder/spell/invoked/convergence/cast(list/targets, mob/living/user)
-	. = ..()
-	if(isliving(targets[1]))
-		var/mob/living/target = targets[1]
-		target.visible_message(span_info("A convergence of fates surrounds [target]!"), span_notice("My past and present converge as one!"))
-		if(iscarbon(target))
-			var/mob/living/carbon/C = target
-			C.apply_status_effect(/datum/status_effect/buff/convergence)
-			C.apply_status_effect(/datum/status_effect/buff/fortify)
-		else
-			target.adjustBruteLoss(-50)
-			target.adjustFireLoss(-50)
-		return TRUE
-	revert_cast()
-	return FALSE
-//Caustic Edit End
